emitter 리팩토링
