+++ data/board/__init__.py
@@ -2,4 +2,3 @@ from .internal.point import Point, PointRange, is_overlap, get_overlap
 from .internal.tile import Tile
 from .internal.tiles import Tiles
 from .internal.section import Section, abs_to_sec, SectionFlag
-from .internal import tile_emitter  # emitter 핸들러 자동 등록 보장
