+++ data/cursor/__init__.py
@@ -1,4 +1,3 @@
 from .internal.cursor import Cursor
 from .internal.rankrange import RankRange, CursorRankRange
 from .internal.items import Items, ItemType
-from .internal import event_emitter  # emitter 핸들러 자동 등록 보장
