+++ data/cursor/emitter/__init__.py
@@ -1,0 +1,2 @@
+from .internal.emitter import CursorEmitter
+from .internal import handlers  # 핸들러 자동 등록 보장
