+++ data/cursor/internal/cursor.py
@@ -68,8 +68,3 @@ class Cursor(DataObj):
             score=score,
             items=items.copy()
         )
-
-
-# CursorEmitter를 클래스 속성으로 할당 (dataclass fields에 포함되지 않도록)
-from .event_emitter import CursorEmitter
-Cursor.emitter = CursorEmitter
